Add blog picker to gutenberg editor
Also:
- refactor PostEditorUtil as PublishablePostEditor extension
- Extract PublishablePostEditor extensions to their own files
